Cricket Game.
This is a simple JavaScript-based project that contains three choices BAT, BALL AND STAMP and two users one is the computer that randomly chooses any choice and the other one is you. you can also choose one of the choices and according to the rules you either win or loss and finally players score can be recorded.